business-assistant-platform/backend/requirements.txt
# File: backend/requirements.txt
# Purpose: Python dependencies for the backend scaffold and future layer implementation.

fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.42
psycopg2-binary==2.9.10
pymongo==4.13.2
redis==6.2.0
celery==5.5.3
qdrant-client==1.15.1
minio==7.2.16
pypdf==5.9.0
python-docx==1.2.0
pandas==2.3.1
numpy==2.3.2
scikit-learn==1.7.1
xgboost==3.0.4
mlflow==3.1.4
httpx==0.28.1
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.28
